jhipster-starter/jhipster-vuejs-starter/src/main/webapp/app/entities/location/location.service.ts
import axios from 'axios';

import { ILocation } from '@/shared/model/location.model';

const baseApiUrl = 'api/locations';

export default class LocationService {
  public find(id): Promise<ILocation> {
    return new Promise<ILocation>(resolve => {
      axios.get(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public retrieve(paginationQuery?: any): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(baseApiUrl).then(function(res) {
        resolve(res);
      });
    });
  }

  public delete(id): Promise<any> {
    return new Promise<any>(resolve => {
      axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res);
      });
    });
  }

  public create(entity): Promise<ILocation> {
    return new Promise<ILocation>(resolve => {
      axios.post(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public update(entity): Promise<ILocation> {
    return new Promise<ILocation>(resolve => {
      axios.put(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }
}
